@@ -3,7 +3,7 @@ requires = [ "poetry-core>=1.0.0",]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-version = "0.1.9"
+version = "0.2.0"
 name = "destination-sqlite"
 description = "Destination implementation for Sqlite."
 authors = [ "Airbyte <contact@airbyte.io>",]
@@ -16,8 +16,8 @@ repository = "https://github.com/airbytehq/airbyte"
 include = "destination_sqlite"
 
 [tool.poetry.dependencies]
-python = "^3.9,<3.12"
-airbyte-cdk = "==0.1.72"
+python = "^3.10,<3.13"
+airbyte-cdk = "^6.36.4"
 
 [tool.poetry.scripts]
 destination-sqlite = "destination_sqlite.run:run"
