@@ -15,6 +15,7 @@ import io.airbyte.cdk.load.message.DestinationFile
 import io.airbyte.cdk.load.message.DestinationFileStreamComplete
 import io.airbyte.cdk.load.message.DestinationFileStreamIncomplete
 import io.airbyte.cdk.load.message.DestinationRecord
+import io.airbyte.cdk.load.message.DestinationRecordAirbyteValue
 import io.airbyte.cdk.load.message.DestinationRecordStreamComplete
 import io.airbyte.cdk.load.message.DestinationRecordStreamIncomplete
 import io.airbyte.cdk.load.message.DestinationStreamAffinedMessage
@@ -23,13 +24,20 @@ import io.airbyte.cdk.load.message.GlobalCheckpoint
 import io.airbyte.cdk.load.message.GlobalCheckpointWrapped
 import io.airbyte.cdk.load.message.MessageQueue
 import io.airbyte.cdk.load.message.MessageQueueSupplier
+import io.airbyte.cdk.load.message.PartitionedQueue
+import io.airbyte.cdk.load.message.PipelineEndOfStream
+import io.airbyte.cdk.load.message.PipelineEvent
+import io.airbyte.cdk.load.message.PipelineMessage
 import io.airbyte.cdk.load.message.QueueWriter
 import io.airbyte.cdk.load.message.SimpleBatch
 import io.airbyte.cdk.load.message.StreamCheckpoint
 import io.airbyte.cdk.load.message.StreamCheckpointWrapped
 import io.airbyte.cdk.load.message.StreamEndEvent
+import io.airbyte.cdk.load.message.StreamKey
 import io.airbyte.cdk.load.message.StreamRecordEvent
 import io.airbyte.cdk.load.message.Undefined
+import io.airbyte.cdk.load.pipeline.InputPartitioner
+import io.airbyte.cdk.load.pipeline.LoadPipeline
 import io.airbyte.cdk.load.state.Reserved
 import io.airbyte.cdk.load.state.SyncManager
 import io.airbyte.cdk.load.task.DestinationTaskLauncher
@@ -42,6 +50,7 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Named
 import jakarta.inject.Singleton
+import java.util.concurrent.ConcurrentHashMap
 
 interface InputConsumerTask : Task
 
@@ -62,15 +71,25 @@ class DefaultInputConsumerTask(
     private val inputFlow: ReservingDeserializingInputFlow,
     private val recordQueueSupplier:
         MessageQueueSupplier<DestinationStream.Descriptor, Reserved<DestinationStreamEvent>>,
-    private val checkpointQueue: QueueWriter<Reserved<CheckpointMessageWrapped>>,
+    val checkpointQueue: QueueWriter<Reserved<CheckpointMessageWrapped>>,
     private val syncManager: SyncManager,
     private val destinationTaskLauncher: DestinationTaskLauncher,
     @Named("fileMessageQueue")
     private val fileTransferQueue: MessageQueue<FileTransferQueueMessage>,
+
+    // Required by new interface
+    @Named("recordQueue")
+    private val recordQueueForPipeline:
+        PartitionedQueue<Reserved<PipelineEvent<StreamKey, DestinationRecordAirbyteValue>>>,
+    private val loadPipeline: LoadPipeline? = null,
+    private val partitioner: InputPartitioner,
+    private val openStreamQueue: QueueWriter<DestinationStream>
 ) : InputConsumerTask {
     private val log = KotlinLogging.logger {}
 
     override val terminalCondition: TerminalCondition = OnSyncFailureOnly
+
+    private val unopenedStreams = ConcurrentHashMap(catalog.streams.associateBy { it.descriptor })
 
     private suspend fun handleRecord(
         reserved: Reserved<DestinationStreamAffinedMessage>,
@@ -123,6 +142,65 @@ class DefaultInputConsumerTask(
         }
     }
 
+    private suspend fun handleRecordForPipeline(
+        reserved: Reserved<DestinationStreamAffinedMessage>,
+    ) {
+        val stream = reserved.value.stream
+        unopenedStreams.remove(stream)?.let {
+            log.info { "Saw first record for stream $stream; initializing" }
+            // Note, since we're not spilling to disk, there is nothing to do with
+            // any records before initialization is complete, so we'll wait here
+            // for it to finish.
+            openStreamQueue.publish(it)
+            syncManager.getOrAwaitStreamLoader(stream)
+            log.info { "Initialization for stream $stream complete" }
+        }
+        val manager = syncManager.getStreamManager(stream)
+        when (val message = reserved.value) {
+            is DestinationRecord -> {
+                val record = message.asRecordMarshaledToAirbyteValue()
+                manager.incrementReadCount()
+                val pipelineMessage =
+                    PipelineMessage(
+                        mapOf(manager.getCurrentCheckpointId() to 1),
+                        StreamKey(stream),
+                        record
+                    )
+                val partition = partitioner.getPartition(record, recordQueueForPipeline.partitions)
+                recordQueueForPipeline.publish(reserved.replace(pipelineMessage), partition)
+            }
+            is DestinationRecordStreamComplete -> {
+                manager.markEndOfStream(true)
+                log.info { "Read COMPLETE for stream $stream" }
+                recordQueueForPipeline.broadcast(reserved.replace(PipelineEndOfStream(stream)))
+                reserved.release()
+            }
+            is DestinationRecordStreamIncomplete -> {
+                manager.markEndOfStream(false)
+                log.info { "Read INCOMPLETE for stream $stream" }
+                recordQueueForPipeline.broadcast(reserved.replace(PipelineEndOfStream(stream)))
+                reserved.release()
+            }
+            is DestinationFile -> {
+                val index = manager.incrementReadCount()
+                // destinationTaskLauncher.handleFile(stream, message, index)
+                fileTransferQueue.publish(FileTransferQueueMessage(stream, message, index))
+            }
+            is DestinationFileStreamComplete -> {
+                reserved.release() // safe because multiple calls conflate
+                manager.markEndOfStream(true)
+                val envelope =
+                    BatchEnvelope(
+                        SimpleBatch(Batch.State.COMPLETE),
+                        streamDescriptor = message.stream,
+                    )
+                destinationTaskLauncher.handleNewBatch(stream, envelope)
+            }
+            is DestinationFileStreamIncomplete ->
+                throw IllegalStateException("File stream $stream failed upstream, cannot continue.")
+        }
+    }
+
     private suspend fun handleCheckpoint(
         reservation: Reserved<CheckpointMessage>,
         sizeBytes: Long
@@ -135,12 +213,19 @@ class DefaultInputConsumerTask(
             is StreamCheckpoint -> {
                 val stream = checkpoint.checkpoint.stream
                 val manager = syncManager.getStreamManager(stream)
+                val checkpointId = manager.getCurrentCheckpointId()
                 val (currentIndex, countSinceLast) = manager.markCheckpoint()
+                val indexOrId =
+                    if (loadPipeline == null) {
+                        currentIndex
+                    } else {
+                        checkpointId.id.toLong()
+                    }
                 val messageWithCount =
                     checkpoint.withDestinationStats(CheckpointMessage.Stats(countSinceLast))
                 checkpointQueue.publish(
                     reservation.replace(
-                        StreamCheckpointWrapped(sizeBytes, stream, currentIndex, messageWithCount)
+                        StreamCheckpointWrapped(sizeBytes, stream, indexOrId, messageWithCount)
                     )
                 )
             }
@@ -153,8 +238,15 @@ class DefaultInputConsumerTask(
                 val streamWithIndexAndCount =
                     catalog.streams.map { stream ->
                         val manager = syncManager.getStreamManager(stream.descriptor)
+                        val checkpointId = manager.getCurrentCheckpointId()
                         val (currentIndex, countSinceLast) = manager.markCheckpoint()
-                        Triple(stream, currentIndex, countSinceLast)
+                        val indexOrId =
+                            if (loadPipeline == null) {
+                                currentIndex
+                            } else {
+                                checkpointId.id.toLong()
+                            }
+                        Triple(stream, indexOrId, countSinceLast)
                     }
                 val totalCount = streamWithIndexAndCount.sumOf { it.third }
                 val messageWithCount =
@@ -181,8 +273,13 @@ class DefaultInputConsumerTask(
                 inputFlow.collect { (sizeBytes, reserved) ->
                     when (val message = reserved.value) {
                         /* If the input message represents a record. */
-                        is DestinationStreamAffinedMessage ->
-                            handleRecord(reserved.replace(message), sizeBytes)
+                        is DestinationStreamAffinedMessage -> {
+                            if (loadPipeline != null) {
+                                handleRecordForPipeline(reserved.replace(message))
+                            } else {
+                                handleRecord(reserved.replace(message), sizeBytes)
+                            }
+                        }
                         is CheckpointMessage ->
                             handleCheckpoint(reserved.replace(message), sizeBytes)
                         is Undefined -> {
@@ -196,6 +293,7 @@ class DefaultInputConsumerTask(
             log.info { "Closing record queues" }
             catalog.streams.forEach { recordQueueSupplier.get(it.descriptor).close() }
             fileTransferQueue.close()
+            recordQueueForPipeline.close()
         }
     }
 }
@@ -208,14 +306,22 @@ interface InputConsumerTaskFactory {
             MessageQueueSupplier<DestinationStream.Descriptor, Reserved<DestinationStreamEvent>>,
         checkpointQueue: QueueWriter<Reserved<CheckpointMessageWrapped>>,
         destinationTaskLauncher: DestinationTaskLauncher,
-        fileTransferQueue: MessageQueue<FileTransferQueueMessage>
+        fileTransferQueue: MessageQueue<FileTransferQueueMessage>,
+
+        // Required by new interface
+        recordQueueForPipeline:
+            PartitionedQueue<Reserved<PipelineEvent<StreamKey, DestinationRecordAirbyteValue>>>,
+        loadPipeline: LoadPipeline?,
+        partitioner: InputPartitioner,
+        openStreamQueue: QueueWriter<DestinationStream>,
     ): InputConsumerTask
 }
 
 @Singleton
 @Secondary
-class DefaultInputConsumerTaskFactory(private val syncManager: SyncManager) :
-    InputConsumerTaskFactory {
+class DefaultInputConsumerTaskFactory(
+    private val syncManager: SyncManager,
+) : InputConsumerTaskFactory {
     override fun make(
         catalog: DestinationCatalog,
         inputFlow: ReservingDeserializingInputFlow,
@@ -224,6 +330,13 @@ class DefaultInputConsumerTaskFactory(private val syncManager: SyncManager) :
         checkpointQueue: QueueWriter<Reserved<CheckpointMessageWrapped>>,
         destinationTaskLauncher: DestinationTaskLauncher,
         fileTransferQueue: MessageQueue<FileTransferQueueMessage>,
+
+        // Required by new interface
+        recordQueueForPipeline:
+            PartitionedQueue<Reserved<PipelineEvent<StreamKey, DestinationRecordAirbyteValue>>>,
+        loadPipeline: LoadPipeline?,
+        partitioner: InputPartitioner,
+        openStreamQueue: QueueWriter<DestinationStream>,
     ): InputConsumerTask {
         return DefaultInputConsumerTask(
             catalog,
@@ -233,6 +346,12 @@ class DefaultInputConsumerTaskFactory(private val syncManager: SyncManager) :
             syncManager,
             destinationTaskLauncher,
             fileTransferQueue,
+
+            // Required by new interface
+            recordQueueForPipeline,
+            loadPipeline,
+            partitioner,
+            openStreamQueue,
         )
     }
 }
