@@ -13,22 +13,11 @@ import jakarta.inject.Singleton
 @Singleton
 @Requires(env = ["MockPathFactory"])
 open class MockPathFactory : PathFactory {
-    open var doSupportStaging = false
-
-    override val supportsStaging: Boolean
-        get() = doSupportStaging
     override val finalPrefix: String
         get() = "prefix"
 
     private fun fromStream(stream: DestinationStream): String {
         return "${stream.descriptor.namespace}/${stream.descriptor.name}"
-    }
-
-    override fun getStagingDirectory(
-        stream: DestinationStream,
-        substituteStreamAndNamespaceOnly: Boolean
-    ): String {
-        return "$finalPrefix/staging/${fromStream(stream)}"
     }
 
     override fun getFinalDirectory(
@@ -41,22 +30,16 @@ open class MockPathFactory : PathFactory {
     override fun getPathToFile(
         stream: DestinationStream,
         partNumber: Long?,
-        isStaging: Boolean,
         extension: String?
     ): String {
-        val prefix = if (isStaging) getStagingDirectory(stream) else getFinalDirectory(stream)
+        val prefix = getFinalDirectory(stream)
         return "${prefix}file"
     }
 
     override fun getLongestStreamConstantPrefix(
         stream: DestinationStream,
-        isStaging: Boolean
     ): String {
-        return if (isStaging) {
-            getStagingDirectory(stream)
-        } else {
-            getFinalDirectory(stream)
-        }
+        return getFinalDirectory(stream)
     }
 
     override fun getPathMatcher(
