@@ -3,7 +3,7 @@
 from unittest import TestCase
 from unittest.mock import Mock
 
-from airbyte_protocol.models import AirbyteCatalog, AirbyteStream, SyncMode
+from airbyte_cdk.models import AirbyteCatalog, AirbyteStream, SyncMode
 
 from erd.dbml_assembler import DbmlAssembler, Source
 from tests.builder import RelationshipBuilder
