@@ -14,7 +14,7 @@ class S3DataLakePerformanceTest :
     BasicPerformanceTest(
         configContents = Files.readString(S3DataLakeTestUtil.GLUE_CONFIG_PATH),
         configSpecClass = S3DataLakeSpecification::class.java,
-        defaultRecordsToInsert = 500_000,
+        defaultRecordsToInsert = 1_000_000,
         micronautProperties =
             S3DataLakeTestUtil.getAwsAssumeRoleCredentials().asMicronautProperties(),
     )
