@@ -198,7 +198,7 @@ const config = {
       prism: {
         theme: lightCodeTheme,
         darkTheme: darkCodeTheme,
-        additionalLanguages: ["bash", "diff", "json"],
+        additionalLanguages: ["bash", "diff", "json", "hcl"],
       },
     }),
 };
