@@ -745,7 +745,6 @@ module.exports = {
       id: "using-airbyte/pyairbyte/getting-started",
     },
     understandingAirbyte,
-    contributeToAirbyte,
     {
       type: "category",
       label: "Licenses",
@@ -761,6 +760,7 @@ module.exports = {
       ],
     },
     sectionHeader("Community"),
+    contributeToAirbyte,
     "community/getting-support",
     "community/code-of-conduct",
     sectionHeader("Product Updates"),
