@@ -5,16 +5,20 @@
 import json
 import os
 
-from airbyte_cdk.models import AirbyteMessage, ConfiguredAirbyteCatalog, ConfiguredAirbyteStream, DestinationSyncMode
+from airbyte_cdk.models import ConfiguredAirbyteCatalog, ConfiguredAirbyteStream, DestinationSyncMode
 
 
 def configure_catalog():
-    record = AirbyteMessage.parse_raw(input())
-    for stream in record.catalog.streams:
-        stream.json_schema = {}
+    record_dict = json.loads(input())
+    catalog_streams = record_dict.get("catalog", {}).get("streams", [])
+    for stream in catalog_streams:
+        stream["json_schema"] = {}
     streams = [
-        ConfiguredAirbyteStream(stream=stream, sync_mode=stream.supported_sync_modes[0], destination_sync_mode=DestinationSyncMode.append)
-        for stream in record.catalog.streams
+        ConfiguredAirbyteStream(
+            stream=stream.get("name"), sync_mode=stream.get("supported_sync_modes", [])[0], destination_sync_mode=DestinationSyncMode.append
+        )
+        for stream in catalog_streams
+        if stream.get("supported_sync_modes")
     ]
     configured_catalog = ConfiguredAirbyteCatalog(streams=streams)
 
@@ -23,4 +27,17 @@ def configure_catalog():
         os.mkdir(default_folder)
     output_file_name = os.path.join(default_folder, "configured_catalog.json")
     with open(output_file_name, "w") as outfile:
-        json.dump(json.loads(configured_catalog.json()), outfile, indent=2, sort_keys=True)
+        # Create a dictionary representation of the configured catalog
+        result = {"streams": []}
+        for stream in configured_catalog.streams:
+            stream_dict = {
+                "stream": {
+                    "name": stream.stream.name if hasattr(stream.stream, "name") else stream.stream,
+                    "supported_sync_modes": ["full_refresh"],
+                    "json_schema": {},
+                },
+                "sync_mode": str(stream.sync_mode),
+                "destination_sync_mode": "append",
+            }
+            result["streams"].append(stream_dict)
+        json.dump(result, outfile, indent=2, sort_keys=True)
